push to docker test: don't get fooled by podman
Gating tests have failed (at least) twice already because
one of the push.bats tests runs "which docker" and skips
if it's missing. Sadly, some gating-test systems install
podman-docker (possibly when there's a bodhi that combines
podman and buildah). This causes the test to fail.

Solution: confirm that if docker exists, it isn't podman
in disguise. Since we need to do a similar check in
the pull-from-docker-daemon test in pull.bats, refactor
that into a common skip_if_no_docker helper.

Signed-off-by: Ed Santiago <santiago@redhat.com>