imagebuildah.StageExecutor.volumeCacheSaveVFS(): set up bind mounts
When building with non-overlay drivers, we archive the contents of
volumes before RUN instructions and restore their contents when the RUN
instruction is finished.

The buildah.Run() ("buildah run") logic defaults to creating a copy of
the contents of built-in volumes and bind-mounting them when running
subcommands, but does not restore the contents afterward because the
rootfs's contents aren't changed.  The rootfs's contents do get changed
by buildah.Add() ("buildah add/copy"), and the volumes go out of sync
with the rootfs because we don't really know what to do then.

The save/restore logic in imagebuildah for non-overlay paths wasn't
adding mounts to the list used by buildah.Run(), so it was continuing to
set up bind mounts even in cases where our cache was invalidated, which
produced this kind of loss of synchronization.  The overlay path wasn't
affected because it adds overlay mounts to the list we pass to
buildah.Run(), and buildah.Run() doesn't override mount points.  So to
avoid this for the VFS methods, add bind mounts to the rootfs, which
we're updating anyway.  They're basically redundant, but they work
around the problem.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>