run.bats: fix flake in run-user test
The run-user test creates a test user with random UID and GID.
If the randomly-picked UID matches that of an existing account,
the test will flake because the account's default group will
almost certainly not match the one we generated.

Solution: only pick UIDs and GIDs that are not already present
in the container image.

Signed-off-by: Ed Santiago <santiago@redhat.com>