Multi-arch github-action workflow unification
This is a port from the podman repository, of substantially the same
workflow with a number of bugfixes and readability improvements
compared to the original.  Same for the README.md updates.

The significant changes compared to the prior implementation are:

* Run periodically instead of only after every master push.
* Add a build for the "testing" image flavor.
* Fix a blank `org.opencontainers.image.source` value.
* Instead of pushing a `main` (or `master`) tagged image, use `latest`.
* Simplify use of env. vars. and workflow vars.

Note: Aside from a `s/podman/buildah/g` this commit makes the
buildah and podman workflows identical.  This is needed to better
support a smooth transition to a future/intended unification effort.

In other words, I intend to develop a single, shared workflow/script
that can be used for all three: skopeo, buildah, and podman.

Signed-off-by: Chris Evich <cevich@redhat.com>