Notification email for cirrus-cron build failures
This duplicates the github-action workflow of the same name/purpose from
the podman reposiory.  Every day it will check the most recent list
of Cirrus-cron initiated builds, and only send an e-mail if any failed.

Some measures are taken in this version to re-use a script and some
data from the podman repo.  There is no way (as of this commit) to
avoid duplicating the workflow file.

Signed-off-by: Chris Evich <cevich@redhat.com>