Fix copy race while walking paths
During a copy operation which descends through a directory tree,
It's possible for a referenced file to become inaccessible (by unlink
or permission change or whatever).  During the walk of paths to copy,
an `Lstat()` is run on each item, and any error passed into the handler
function to deal with.  Subsequently, if there is no error, the file
is examined for inclusion/exclusion by the handler.

Unfortunately, this introduces a TOCTOU race condition for files which
become inaccessible even if they would otherwise be excluded.  For
example a file or directory under /proc or /sys (which frequently and
unpredictably change).  This was the original cause encountered during
podman integration testing.

It's impractical to actually fix this race at the file-level, without
introducing negative effects to any source-container operations.  It's
also questionably useful to offer a command-line option to offload the
choice to the user.  Instead, follow the behavior of the `tar` command
for this situation: Issue a warning to the user, and ignore the
problematic item (don't copy it).

Also add a test resembling the podman test which originally caught this
race.  While not reliable, it does introduce a non-zero chance of
hitting the race condition - and handling the new warning properly.

Signed-off-by: Chris Evich <cevich@redhat.com>