buildah bud: fix containers/podman/issues/10307
Fix a regression in `buildah bud`.  In case of a cache hit or an
intermediate image, the pull policy should be set to "never" so that we
enforce a local-only lookup without even attempting to reach out to a
registry.

The regression was detected in Podman which defaults to --layers=true
which ultimately broke when setting --pull or --pull-always.  In case of
a cache hit, the image reference refers to a local image which conflicts
with an "always" pull policy.

Also extend an existing bud test to make sure we don't regress on it in
the future.

Signed-off-by: Valentin Rothberg <rothberg@redhat.com>