Multi-arch: Always push updated version-tagged img
Fixes #3365

Given a near daily release of Fedora updates, some addressing security
issues, it's safest for users to always have an updated version-tagged
image.  The cost of this is more image churn, but for a relatively small
image this should hopefully be a minor burden vs improved security.

Note: This is not a perfect solution, as there may be no
hard-dependencies between buildah and other packages with security
updates.  However it is an improvement over the previous regime where we
simply pushed the latest version-tagged image once, then never touched
it again.

Minor: Rename `gen_fqin` identifier to `reponame_reg` to be more
consistent with it's sister block `containers_reg`.

Signed-off-by: Chris Evich <cevich@redhat.com>