Bump go for vendor-in-container from 1.13 to 1.16
While vendoring c/storage v1.34.0 I was having issues
with io/fs not being able to be vendored via calls in the
latest BurntSushi.  Fortunately I remembered a change to
the Skopeo Makefile that @vrothberg had made recently and
viola, that fixed it.

Also per @nalind's review, I've bumped Go from 1.13 to 1.6 in the
.cirrus.yml file.

[NO NEW TESTS NEEDED]

Signed-off-by: TomSweeneyRedHat <tsweeney@redhat.com>