executor: make sure imageMap is updated with terminatedStage
When the terminatedStage map has its entry for a stage set, make sure
that we set the corresponding entry in the imageMap while holding the
mutex for it, eliminating the sliver of time when the first is set but
the second isn't, which could bite StageExecutor.Execute(), which waits
for the first and then reads the second.

Make terminatedStage record the error if a stage doesn't complete
successfully, and have waitForStage() return that error if it's set.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>