Makefile: update cross targets
Update the "cross" makefile target to just try building on every Linux,
Darwin, or Windows architecture that the current version of Go supports.

That way, the next time we upgrade Go to a version that supports more
architectures, we won't have to remember to update the list of
cross-compile targets that we try to build in CI to catch
cross-compilation problems.

Marking this as not needing new tests because it directly modifies one
of our CI tests.
[NO NEW TESTS NEEDED]

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>