unmarshalConvertedConfig(): handle zstd compression
The Docker manifest format doesn't currently support listing layers
compressed using zstd, so we trigger an error when we try to convert an
in-memory OCI manifest to the Docker format as a preliminary step in
reading the image's config blob in the Docker format.

Instead, first create a temporary copy of the manifest, and then force
the MIME types for all layers in the temporary copy of the manifest to
appear to be compressed using gzip.  Both OCI and Docker formats will
accept the resulting manifest without issue.  We throw the copy away
after we've read the config blob, so the wackiness should be contained.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>
(cherry picked from commit 9073afddef701846acbf8f2db3962573f54e851f)