Makefile: check for `-race` using `-mod=vendor`
When checking if `go test` supports `-race` on the current platform, use
`-mod=vendor` if it's available, so that the compiler won't try to pull
down our dependencies, and run it against a package that has fewer
dependencies.

Run `go tool dist list` once, and filter its results to derive the
cross-compile target lists, rather than running it three times.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>