docs/tutorials: update
Update tutorials to not expect `buildah run` to do anything if it isn't
given a command to run.  In some cases (including when we need to listen
for incoming connections when we might well not be root) this means we
use `podman run` instead.

Try to avoid using the terms container and image as though they're
interchangeable, which just creates confusion.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>