bud --base-images
Add a new flag `--base-images` to `buildah bud` which allows for
extracting the base images without performing any build.

Extracting the base images *before* a build is needed in `podman-remote
build` to properly resolve short names prior to performing the build.

Podman will consume the feature via the Go code but exposing it on
the `bud` CLI allows for easier testing.

Signed-off-by: Valentin Rothberg <rothberg@redhat.com>