setup the netns in the buildah parent process
Do not configure the netns in the runtime child process, this removes
the need to send the network options to the child. This will be needed
for the new libnetwork network interface which cannot be transfered as
json.

To synchronize this between the child and parent we use two pipe pairs.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>