Fix platform handling for empty os/arch values
If you run `buildah bud --platform windows/` the os is ignored at the
moment and it uses the host os. We should still use the os from the
platform in this case. This fixes an issue with podman-remote build
where it is possible that only the os is set in the platform string.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>