Cirrus: Run int. tests in parallel with unit
This is somewhat of a band-aid since these two tasks take 45+ minutes
each.  Running them in parallel improves the overall runtime when
tests are otherwise/generally passing.  If/when unit tests fail, the
status won't update until integration tasks complete, but the failure
will be immediately observable in a PR.  So it's a minor price to pay
for improved runtime in the general case.

Signed-off-by: Chris Evich <cevich@redhat.com>