imagebuildah.BuildDockerfiles(): create the jobs semaphore
Make sure that, if we're given neither a concurrent jobs count nor a
semaphore to use, we create a semaphore.  Otherwise, each platform in a
multi-platform build will run with maximum concurrency, which isn't the
desired default.  Our CLI always passes in a Jobs count, so it hasn't
been affected by this problem.

[NO NEW TESTS NEEDED]

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>