speed up combination-namespaces test
Commit d05957a6f6de added the private value for namespace flags and kept
container as alias for backwards compat. Commit b480ce832f3e added the
cgroup falg so the test has 3^6 combinations to test. This takes way to
long. Since container is the same as private we only test private. The
container value is already covered in the other namespace tests, see
general_namespace() function.

With this commit the test time went down from 23 min to 3 min on my
laptop.

Fixes #3768

Signed-off-by: Paul Holzinger <pholzing@redhat.com>