tests/copy: read correct containers.conf
`DefaultConfig()` from c/common/pkg/config should not be used here.
This only contains the built in defaults but not the actual values from
the files on disk. The correct way is to call `Default()` and since this
already handles the `CONTAINERS_CONF` env var we do not need to handle
it manually.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>