run: set parent-death signals and forward SIGHUP/SIGINT/SIGTERM
Restore setting of the parent-death signal when we're running
subprocesses in Run(), so that if we get killed, the child processes
will also get killed.

While a child process is running, if we receive SIGHUP, SIGINT, or
SIGTERM, forward the signal to our child process unless it's the command
we're executing, which we SIGKILL without mercy, and finish the current
routine, which will then notice that the child process has exited and
return an error to its caller.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>