Create shorter names for containers based on image IDs
When we specify the image name to use for creating a container by using
its ID, truncate the ID before appending the suffix.  An untruncated ID
is already as long as the static HOST_NAME_MAX value used by some
applications, and we started making the container's name the FQDN of the
container in f8c152694c1edce4b9b1d80d926f94226caab90c, which would have
created problems for those applications.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>