oci spec: change special mount points for namespaces
The current logic in setupRootlessSpecChanges has quite a few problems.
First of this logic is a noop for rootful and when we have a netns.
However pid and ipc could still need special mounts.

In Podman we have already working code for this so this reuses the same
logic and adapts it to buildah.
https://github.com/containers/podman/blob/4ba71f955a944790edda6e007e6d074009d437a7/pkg/specgen/generate/oci.go#L178

Since there is already logic to handle these mounts in buildah we will
add it there and also handle the special rootless case there.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>