Tests: disable three gating-test failures
One is a certificate error:

    $ buildah push --tls-verify=true alpine localhost:5000/my-alpine
    Get "https://localhost:5000/v2/": x509: certificate relies on legacy Common Name field, use SANs instead

This was addressed in #3014, but back then it failed with "use
SANs or set a magic DEBUG envariable"; that envariable no longer
works.

The other two errors are due to github no longer supporting git:// URLs.
This was addressed in #3701, but that's too hard to backport, so let's
just skip the failing tests.

Signed-off-by: Ed Santiago <santiago@redhat.com>