Initialize network backend before first pull
After clean install, it is necessary to decide the network backend
before any image is pulled so that `netavark` is chosen correctly.
Without this change, if `buildah pull` is executed at first, the network
backend is not determined and an image is pulled. This results in
choosing `cni` at a next command because an image already exists while
`netavark` is chosen if `buildah from` or `buildah bud` is called at
first.

Signed-off-by: Hironori Shiina <shiina.hironori@jp.fujitsu.com>