copier.Put(): write to read-only directories
Try to improve our ability to write to directories that aren't
writable.  If we encounter an EPERM error while attempting to create an
item, attempt to temporarily make writable the directory that we're
writing the item to, and restore its permissions on our way out.

The error usually isn't seen when run as UID 0, whether in a user
namespace or not, which is usually how we're called, but running the
unit tests as an unprivileged user will verify it.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>