buildah build: --annotation and --label should remove values
Have the handling of the `--annotation` and `--label` options for
`buildah build` explicitly remove values when they're not given an "="
before an empty value, instead of setting them to empty values.  The
final result doesn't seem to care about the distinction.

Allow the value of the `io.buildah.version` label to be manually
specified on the `buildah build` command line if someone really wants to
override the value that we would otherwise set by default.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>