Fix up buildah images
Fedora no longer ships /etc/containers/storage.conf but
/usr/share/containers/storage.conf, this causes the buildimages
to fail to build. This PR uses the storage.conf in /usr.

Removed Centos7 image, since this distribution is no longer supported
and no reason to ship an image specific to this distribution.
Removed stablebyhand image, since this does not look like it is needed
any longer, and was not being maintained.

[NO NEW TESTS NEEDED]

Signed-off-by: Daniel J Walsh <dwalsh@redhat.com>