Remove redundant buildahimages-are-sane validation
This has been replaced with a more comprehensive multi-arch test which
is activated by including `[CI:BUILD]` in the PR title.  It exercises
not only the images, but buildah and the entire build_push toolchain.
Further, these images are built daily by Cirrus-Cron and monitored for
any failures making testing in every PR much less useful.

Signed-off-by: Chris Evich <cevich@redhat.com>